+++ src/runner.rs
@@ -5,8 +5,11 @@ use rusoto_qldb_session::{QldbSession, QldbSessionClient};
 use rustyline::error::ReadlineError;
 use tracing::instrument;
 
-use crate::settings::{Environment, ExecuteStatementOpt};
-use crate::transaction::ShellTransaction;
+use crate::{
+    command,
+    settings::{Environment, ExecuteStatementOpt},
+};
+use crate::{settings::Setter, transaction::ShellTransaction};
 use crate::{Deps, QldbShellError};
 
 pub(crate) enum ProgramFlow {
@@ -63,12 +66,14 @@ where
     C: QldbSession + Send + Sync + Clone + 'static,
 {
     pub(crate) async fn start(&mut self) -> Result<ProgramFlow> {
-        self.deps.ui.println(
-            r#"Welcome to the Amazon QLDB Shell!
+        if self.deps.env.display_welcome().value {
+            self.deps.ui.println(
+                r#"Welcome to the Amazon QLDB Shell!
 
 To start a transaction type 'start transaction', after which you may enter a series of PartiQL statements.
 When your transaction is complete, enter 'commit' or 'abort' as appropriate."#,
-        );
+            );
+        }
 
         loop {
             match self.tick().await {
@@ -86,7 +91,7 @@ When your transaction is complete, enter 'commit' or 'abort' as appropriate."#,
             None => self
                 .deps
                 .ui
-                .set_prompt(format!("{} ", self.deps.env.prompt.value)),
+                .set_prompt(format!("{} ", self.deps.env.prompt().value)),
             Some(_) => self.deps.ui.set_prompt(format!("qldb *> ")),
         }
 
@@ -106,20 +111,26 @@ When your transaction is complete, enter 'commit' or 'abort' as appropriate."#,
                     }
                 }
             }
-            Err(ReadlineError::Interrupted) => {
-                self.deps.ui.println("CTRL-C");
-                TickFlow::Again
-            }
-            Err(ReadlineError::Eof) => self.handle_break().await?,
-            Err(err) => {
-                self.deps.ui.warn(&format!("Error: {:?}", err));
-                TickFlow::Exit
-            }
+            Err(err) => match err.downcast::<ReadlineError>() {
+                Ok(ReadlineError::Interrupted) => {
+                    if self.deps.env.display_ctrl_signals().value {
+                        self.deps.ui.println("CTRL-C");
+                    }
+                    TickFlow::Again
+                }
+                Ok(ReadlineError::Eof) => self.handle_break().await?,
+                err => {
+                    self.deps.ui.warn(&format!("Error: {:?}", err));
+                    TickFlow::Exit
+                }
+            },
         })
     }
 
     pub(crate) async fn handle_break(&mut self) -> Result<TickFlow> {
-        self.deps.ui.println("CTRL-D");
+        if self.deps.env.display_ctrl_signals().value {
+            self.deps.ui.println("CTRL-D");
+        }
         Ok(if let Some(_) = self.current_transaction {
             self.handle_abort().await?;
             TickFlow::Again
@@ -142,14 +153,45 @@ When your transaction is complete, enter 'commit' or 'abort' as appropriate."#,
             "env" => self.handle_env(),
             "show tables" => self.handle_show_tables().await?,
             "use" => return Ok(TickFlow::Restart), // TODO: implement
-            _ => Err(QldbShellError::UnknownCommand)?,
+            _ => self.handle_complex_command(line)?,
         }
 
         Ok(TickFlow::Again)
     }
 
+    pub(crate) fn handle_complex_command(&mut self, line: &str) -> Result<()> {
+        let iter = line.split_ascii_whitespace();
+        let backslash = match command::backslash(iter) {
+            Ok(b) => b,
+            Err(_) => Err(QldbShellError::UnknownCommand)?,
+        };
+
+        if let command::Backslash::Set(set) = backslash {
+            // FIXME: Hack
+            let mut inner = self.deps.env.inner.lock().unwrap();
+            match set {
+                command::SetCommand::EditMode(ref mode) => {
+                    inner.edit_mode.apply_value(mode, Setter::Environment)
+                }
+                command::SetCommand::TerminatorRequired(ref tf) => {
+                    inner.terminator_required.apply_value(
+                        &match tf {
+                            command::TrueFalse::True => true,
+                            command::TrueFalse::False => false,
+                        },
+                        Setter::Environment,
+                    )
+                }
+            }
+            drop(inner); // FIXME: end hack
+            self.deps.ui.handle_env_set(&set)?;
+        }
+
+        Ok(())
+    }
+
     pub(crate) fn handle_env(&self) {
-        self.deps.ui.println(&format!("{:#?}", self.deps.env));
+        self.deps.ui.println(&format!("{}", self.deps.env));
     }
 
     pub(crate) async fn handle_show_tables(&self) -> Result<()> {