+++ src/rusoto_driver.rs
@@ -1,3 +1,4 @@
+use crate::error;
 use crate::settings::Environment;
 use amazon_qldb_driver::QldbDriverBuilder;
 use amazon_qldb_driver::{retry, QldbDriver};
@@ -9,15 +10,9 @@ use rusoto_core::{
 };
 use rusoto_qldb_session::QldbSessionClient;
 use std::str::FromStr;
-use tracing::warn;
 
 pub async fn build_driver(env: &Environment) -> Result<QldbDriver<QldbSessionClient>> {
-    let provider = profile_provider(&env)?;
-    let region = rusoto_region(&env)?;
-    let creds = match provider {
-        Some(p) => CredentialProvider::Profile(p),
-        None => CredentialProvider::Chain(ChainProvider::new()),
-    };
+    let client = build_rusoto_client(env).await?;
 
     // We disable transaction retries because they don't make sense. Users
     // are entering statements, so if the tx fails they actually have to
@@ -25,11 +20,9 @@ pub async fn build_driver(env: &Environment) -> Result<QldbDriver<QldbSessionCli
     // again, as individual statements may be derived from values seen from
     // yet other statements.
     QldbDriverBuilder::new()
-        .ledger_name(&env.ledger.value)
-        .region(region)
-        .credentials_provider(creds)
+        .ledger_name(env.ledger().value)
         .transaction_retry_policy(retry::never())
-        .build()
+        .build_with_client(client)
         .await
 }
 
@@ -40,11 +33,30 @@ pub(crate) async fn health_check_start_session(env: &Environment) -> Result<()> 
     session_client
         .send_command(SendCommandRequest {
             start_session: Some(StartSessionRequest {
-                ledger_name: env.ledger.value.clone(),
+                ledger_name: env.ledger().value,
             }),
             ..Default::default()
         })
-        .await?;
+        .await
+        .map_err(|e| {
+            error::usage_error(
+                format!(
+                    r#"Unable to connect to ledger `{}`.
+
+Please check the following:
+
+- That you have specified a ledger that exists and is active
+- That the AWS region you are targeting is correct
+- That your AWS credentials are setup
+- That your AWS credentials grant access on this ledger
+
+The following error chain may have more information:
+"#,
+                    env.ledger().value
+                ),
+                e,
+            )
+        })?;
 
     Ok(())
 }
@@ -59,7 +71,8 @@ async fn build_rusoto_client(env: &Environment) -> Result<QldbSessionClient> {
 
     let mut hyper = HttpClient::new()?;
     hyper.local_agent(format!(
-        "QLDB Driver for Rust v{}",
+        "QLDB Driver for Rust v{}/QLDB Shell for Rust v{}",
+        amazon_qldb_driver::version(),
         env!("CARGO_PKG_VERSION")
     ));
 
@@ -88,9 +101,10 @@ impl ProvideAwsCredentials for CredentialProvider {
 }
 
 fn profile_provider(env: &Environment) -> Result<Option<ProfileProvider>> {
-    let it = match &env.profile.value {
+    let it = match env.profile().value {
         Some(p) => {
-            let mut prof = ProfileProvider::new()?;
+            let mut prof = ProfileProvider::new()
+                .map_err(|e| error::usage_error("Unable to create profile provider", e))?;
             prof.set_profile(p);
             Some(prof)
         }
@@ -102,21 +116,18 @@ fn profile_provider(env: &Environment) -> Result<Option<ProfileProvider>> {
 
 // FIXME: Default region should consider what is set in the Profile.
 fn rusoto_region(env: &Environment) -> Result<Region> {
-    let it = match (&env.region.value, &env.qldb_session_endpoint.value) {
+    let it = match (env.region().value, env.qldb_session_endpoint().value) {
         (Some(r), Some(e)) => Region::Custom {
-            name: r.to_owned(),
-            endpoint: e.to_owned(),
+            name: r,
+            endpoint: e,
         },
         (Some(r), None) => match Region::from_str(&r) {
             Ok(it) => it,
-            Err(e) => {
-                warn!("Unknown region {}: {}. If you know the endpoint, you can specify it and try again.", r, e);
-                return Err(e)?;
-            }
+            Err(e) => Err(error::usage_error(format!("Invalid region {}", r), e))?,
         },
         (None, Some(e)) => Region::Custom {
             name: Region::default().name().to_owned(),
-            endpoint: e.to_owned(),
+            endpoint: e,
         },
         (None, None) => Region::default(),
     };