+++ CHANGELOG.md
@@ -12,6 +12,12 @@
      edit_mode = "Emacs" # or Vi
      ```
 
+### Changed
+
+  * When running without a tty, certain UI messages are suppressed.
+     - For example, the welcome message and CTRL-C or CTRL-D is suppressed
+     - This is useful for `echo select * from foo|qldb --ledger example`
+
 ## [2.0.0-alpha6] - 2021-04-28
 
 ### Added