+++ CHANGELOG.md
@@ -19,6 +19,7 @@
      - For example, the welcome message and CTRL-C or CTRL-D is suppressed
      - This is useful for `echo select * from foo|qldb --ledger example`
      - Query metrics are disabled by default
+  * Better error messages on connection fails
 
 ## [2.0.0-alpha6] - 2021-04-28